changed long to float for more accurate measurements

+++ src/main/java/globesort/GlobeSortClient.java
@@ -53,8 +53,8 @@ public class GlobeSortClient {
         IntArray response = serverStub.sortIntegers(request);
         long t4 = System.currentTimeMillis();
         long applicationThroughput = (values.length*1000/(t4-t3));
-        long networkThroughputSecs = (t4-t3-response.getValues(values.length))/(2*1000);
-        long networkThroughput = values.length/networkThroughputSecs;
+        float networkThroughputSecs = (t4-t3-response.getValues(values.length))/(2*1000);
+        float networkThroughput = values.length/networkThroughputSecs;
         System.out.println("Sorted array received. Num Records: " +values.length+ ". Application Throughput(records/sec): " +applicationThroughput);
         System.out.println("Time taken by server to sort(msec): "
                 +response.getValues(values.length)+
